Refactors code to be more readable Follows the single responsibility principle and decoupling

+++ include/string_kata.h
@@ -1,8 +1,13 @@
+// string_kata.h
+// 3/9/2018
+// Ryan Stonebraker
+// Header for the string calculator kata.
+
 #ifndef STRING_KATA_H
 #define STRING_KATA_H
 
 #include <string>
 
-long stringSum (const std::string & inputStr);
+unsigned stringSum (const std::string & inputStr);
 
 #endif
